pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-bigquery = {extras = ["bqstorage", "pandas"], version = "^3.21.0"}
google-cloud-bigquery-storage = "^2.24.0"
pyarrow = "^15.0.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
#future = "^1.0.0"
ibis-framework = {extras = ["bigquery"], version = ">=8.0.0,<9.0.0dev"}
pydata-google-auth = "^1.8.2"
geopandas = "^0.14"
cloudpickle = "^3.0.0"
fsspec = "^2024.3.1"
gcsfs = "^2024.3.1"
pytest = "^8.2.0"
pytest-sugar = "^1.0.0"
google-auth = "^2.29.0"
google-cloud-functions = "^1.16.3"
google-cloud-bigquery-connection = "^1.15.3"
google-cloud-iam = "^2.15.0"
google-cloud-resource-manager = "^1.12.3"
google-cloud-storage = "^2.16.0"
requests = "^2.31.0"
scikit-learn = "^1.4.2"
sqlalchemy = "^2.0.29"
sqlglot = ">=20.8.0,<=20.11"
ipywidgets = "^8.1.2"
humanize = "^4.9.0"
matplotlib = "^3.8.4"
tabulate = "^0.9.0"
networkx = "^3.3"
jellyfish = "^0.8.9"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.7.0"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core=1.8.2"]
build-backend = "poetry.core.masonry.api"
